+++ api/tournaments/admin.py
@@ -1,7 +1,17 @@
 from django.contrib import admin
 from django.utils.translation import gettext_lazy as _
 
+from api.team.models import Team
 from .models import Tournament
+
+
+class TeamsInline(admin.TabularInline):
+    model = Team
+    extra = 0
+    show_change_link = True
+    fields = ('name', 'beachname', 'trainer', 'paid', 'state')
+    ordering = ('state', 'date_signup')
+    classes = ('collapse',)
 
 
 class TournamentAdmin(admin.ModelAdmin):
@@ -18,6 +28,7 @@ class TournamentAdmin(admin.ModelAdmin):
                                    'starting_fee',
                                    'number_of_places')})
     )
+    inlines = [TeamsInline]
 
 
 admin.site.register(Tournament, TournamentAdmin)